+++ CHANGELOG.md
@@ -37,7 +37,7 @@ Version 1.2.2 (not released yet)
 - optimize particles and HUD rendering by reducing number of draw-calls (both in game and level editor)
 - optimize rendering of railgun and rocket launcher scanners by using different back buffer capture technique
 - optimize glare rendering
-- use hardware vertex processing (optimization)
+- use hardware vertex processing in game and level editor (optimization)
 - fix multiple security vulnerabilities
 - stability improvements